lil cleanup
