+++ example-programs/jito-protecc/programs/jito-protecc/Cargo.toml
@@ -16,4 +16,5 @@ cpi = ["no-entrypoint"]
 default = []
 
 [dependencies]
-anchor-lang = "0.27.0"
+anchor-lang = { version = "0.27.0", features = ["init-if-needed"] }
+anchor-spl = {version = "0.27.0", features = ["token"]}
