--- example-programs/jito-protecc/programs/jito-protecc/jito-protecc/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "jito-protecc"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "jito_protecc"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-
-[dependencies]
-anchor-lang = { version = "0.27.0", features = ["init-if-needed"] }
-anchor-spl = {version = "0.27.0", features = ["token"]}
