--- example-programs/jito-protecc/programs/jito-protecc/jito-protecc/Xargo.toml
@@ -1,2 +1,0 @@
-[target.bpfel-unknown-unknown.dependencies.std]
-features = []
